Final push
I’m visualizing women’s rights news coverage according to the night
database. given how many articles there are that year, there is a
ellipse visualization in the background, and also a another that
fluctuates with rosie’s muscle (it’s only really obvious when the
number of articles starts to go over 100). I also tried to display the
headlines of that year in the background. I’m pretty confident I parsed
it correctly to get to the headlines, but when I go to display them,
they cancel out rosie and a bug appears. I commented that bit of the
program out for the sake of it running everything else.